requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
openai==1.35.0
langgraph==0.1.5
jira==3.8.0
requests==2.32.3
